Break out compat index test into a file per function
